add creating post function

+++ controllers/auth.js
@@ -7,7 +7,6 @@ const { isEmpty, isMatched, isEmail, isAdmin } = require("../util/validate");
 module.exports = {
   signUp: async (req, res) => {
     const { username, email, password, confirmPassword } = req.body;
-
     const emptyErrors = isEmpty({ username, email, password, confirmPassword });
     const matchedErrors = isMatched({ password, confirmPassword });
     const emailErrors = isEmail(email);
@@ -52,8 +51,9 @@ module.exports = {
     if (!user || !correctPassword)
       return res.status(404).json({ error: "Incorrect username or password" });
 
+    const userId = user.get("id");
     const role = user.get("role");
-    const token = jwt.sign({ username, role }, process.env.JWT_SECRET, {
+    const token = jwt.sign({ userId, username, role }, process.env.JWT_SECRET, {
       expiresIn: "7d",
     });
 
@@ -64,15 +64,17 @@ module.exports = {
     const { username, email, password, role } = req.body;
     const token = req.get("Authorization");
     const userInfo = jwt_decode(token);
-
     const authorizationErrors = isAdmin(userInfo.role);
     const emptyErrors = isEmpty({ username, email, password, role });
     const emailErrors = isEmail(email);
+    const userFromRequest = await User.findById(userInfo.userId);
 
     if (Object.keys(emptyErrors).length > 0)
       return res.status(400).json(emptyErrors);
     if (Object.keys(emailErrors).length > 0)
       return res.status(400).json(emailErrors);
+    if (!userFromRequest)
+      return res.status(401).json({ error: "Unauthorized" });
     if (Object.keys(authorizationErrors).length > 0)
       return res.status(401).json(authorizationErrors);
 
@@ -99,11 +101,13 @@ module.exports = {
     const { id } = req.params;
     const token = req.get("Authorization");
     const userInfo = jwt_decode(token);
-
+    const userFromRequest = await User.findById(userInfo.userId);
     const authorizationErrors = isAdmin(userInfo.role);
 
     if (Object.keys(authorizationErrors).length > 0)
       return res.status(401).json(authorizationErrors);
+    if (!userFromRequest)
+      return res.status(401).json({ error: "Unauthorized" });
 
     User.findByIdAndDelete(id).then((result) => {
       return res.status(200).json(result);
@@ -115,6 +119,7 @@ module.exports = {
     const { id } = req.params;
     const token = req.get("Authorization");
     const userInfo = jwt_decode(token);
+    const userFromRequest = await User.findById(userInfo.userId);
 
     const emptyErrors = isEmpty({
       username,
@@ -128,9 +133,10 @@ module.exports = {
 
     if (Object.keys(emptyErrors).length > 0)
       return res.status(400).json(emptyErrors);
-
     if (Object.keys(matchedErrors).length > 0)
       return res.status(400).json(matchedErrors);
+    if (!userFromRequest)
+      return res.status(401).json({ error: "Unauthorized" });
 
     const authorizationErrors = isAdmin(userInfo.role);
